#101 [Feature] RecordBasicRepository 추가
